requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
mysqlclient==2.2.4
dj-database-url==0.5.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
